skills/pcsi/requirements.txt
yfinance>=1.2.0
fredapi>=0.5.2
matplotlib>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
PyYAML>=6.0
